# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Marketplace de Servicios API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request a password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordBody"
      responses:
        "200":
          description: Reset email sent (always returns success)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password using token from email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordBody"
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [provider, client, admin, company]
        phone:
          type: ["string", "null"]
        locality:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        cuit:
          type: ["string", "null"]
        companyAddress:
          type: ["string", "null"]
        companyIndustry:
          type: ["string", "null"]
      required:
        - name
        - email
        - password
        - role
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [provider, client, admin, company]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        locality:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        notifyEmail:
          type: boolean
        companyName:
          type: ["string", "null"]
        cuit:
          type: ["string", "null"]
        companyAddress:
          type: ["string", "null"]
        companyIndustry:
          type: ["string", "null"]
        companyApproved:
          type: boolean
        cvUrl:
          type: ["string", "null"]
        cvPublic:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - notifyEmail
        - companyApproved
        - cvPublic
        - createdAt
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
      required:
        - user
...
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        locality:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        notifyEmail:
          type: ["boolean", "null"]
        cvUrl:
          type: ["string", "null"]
        cvPublic:
          type: ["boolean", "null"]
    ForgotPasswordBody:
      type: object
      properties:
        email:
          type: string
      required:
        - email
    ResetPasswordBody:
      type: object
      properties:
        email:
          type: string
        token:
          type: string
        newPassword:
          type: string
      required:
        - email
        - token
        - newPassword